+++ packages/devtools_app/lib/src/screens/memory/panes/chart/data/primitives.dart
@@ -23,3 +23,5 @@ enum ChartInterval {
 
   final String displayName;
 }
+
+const Duration chartUpdateDelay = Duration(milliseconds: 500);