+++ packages/devtools_app/lib/src/screens/memory/shared/primitives/memory_timeline.dart
@@ -9,18 +9,8 @@ import 'package:vm_service/vm_service.dart';
 
 import '../../../../shared/primitives/utils.dart';
 
-enum ContinuesState {
-  none,
-  start,
-  stop,
-  next,
-}
-
 /// All Raw data received from the VM and offline data loaded from a memory log file.
 class MemoryTimeline {
-  static const delayMs = 500;
-  static const Duration updateDelay = Duration(milliseconds: delayMs);
-
   /// This flag will be needed for offline mode implementation.
   final offline = false;
 
@@ -59,8 +49,6 @@ class MemoryTimeline {
 
   /// List of events awaiting to be posted to HeapSample.
   final _eventSamples = <EventSample>[];
-
-  ContinuesState monitorContinuesState = ContinuesState.none;
 
   void postEventSample(EventSample event) {
 /*