+++ packages/devtools_shared/test/server/devtools_extensions_api_test.dart
@@ -39,7 +39,7 @@ void main() {
         host: 'localhost',
         path: ExtensionsApi.apiServeAvailableExtensions,
         queryParameters: {
-          ExtensionsApi.extensionRootPathPropertyName: projectRoot,
+          ExtensionsApi.packageRootUriPropertyName: projectRoot,
         },
       ),
     );
@@ -131,7 +131,7 @@ void main() {
           host: 'localhost',
           path: ExtensionsApi.apiExtensionEnabledState,
           queryParameters: {
-            ExtensionsApi.extensionRootPathPropertyName: projectRoot,
+            ExtensionsApi.packageRootUriPropertyName: projectRoot,
             ExtensionsApi.extensionNamePropertyName: extensionName,
             ExtensionsApi.enabledStatePropertyName: enable.toString(),
           },